# Backend-Code/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Commerce API Docs",
        "description": "Commerce Application Programme Interface",
        "termsOfService": "This API is subject to terms of services and maybe be changed without service notice",
        "contact": {
            "email": "sjo065@uowmail.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": ""
        }
    },
    "servers": [
        {
            "url": "https://http://167.172.62.67/api/v1",
            "description": "Commerce Staging Server"
        },
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
        "/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get all addresses",
                "operationId": "Api/AddressController::index",
                "responses": {
                    "200": {
                        "description": "The resource collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AddressResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Create an address",
                "operationId": "Api/AddressController::store",
                "requestBody": {
                    "description": "The address details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "example": {
                                "name": "John Doe",
                                "email": "john@example.com",
                                "phone": "123-456-7890",
                                "street": "123 Main St",
                                "city": "Springfield",
                                "state": "IL",
                                "country": "US"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The AddressResource.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AddressResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/addresses/:addressId": {
            "patch": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update an address",
                "operationId": "Api/AddressController::update",
                "parameters": [
                    {
                        "name": "addressId",
                        "in": "path",
                        "description": "The resource Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The address details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "example": {
                                "name": "John Doe",
                                "email": "john@example.com",
                                "phone": "123-456-7890",
                                "street": "123 Main St",
                                "city": "Springfield",
                                "state": "IL",
                                "country": "US"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The AddressResource.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AddressResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AddressResource": {
                "properties": {
                    "id": {
                        "description": "The resource Id.",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "The user Id.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the user.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "The phone number of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the user.",
                        "type": "string"
                    },
                    "street": {
                        "description": "The street of the user.",
                        "type": "string"
                    },
                    "city": {
                        "description": "The city of the user.",
                        "type": "string"
                    },
                    "state": {
                        "description": "The state of the user.",
                        "type": "string"
                    },
                    "country": {
                        "description": "The country of the user.",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date the resource was created.",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "The date the resource was updated.",
                        "type": "string"
                    },
                    "full_address": {
                        "description": "The full address of the user.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "400": {
                "description": "Bad Request.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Bad Request.",
                            "errors": []
                        }
                    }
                }
            },
...
            "403": {
                "description": "Forbidden.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Forbidden.",
                            "errors": []
                        }
                    }
                }
            },
            "404": {
                "description": "Resource Not Found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Resource Not Found.",
                            "errors": []
                        }
                    }
                }
            },
...
            "422": {
                "description": "Unprocessable Entity.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Unprocessable Entity.",
                            "errors": []
                        }
                    }
                }
            },
...
            "500": {
                "description": "Server Error.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Server Error.",
                            "errors": []
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/api/v1/login",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}